add option to compile shaders with glslangValidator, not shaderc
if the CMake option USE_SYSTEM_GLSLANG is set, all shader compilation is
done through system calls to the local glslang compiler binary, i.e. the
glslangValidator installed with the Vulkan SDK.

This reduces compile times as shaderc is no longer included.
BUT: you must not use this option when installing Volcanite for
distribution as the shader compiler will be missing on other systems.